% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lnorm3_autofit.R
\name{lnorm3_autofit}
\alias{lnorm3_autofit}
\title{fit lognormal distribution with a shift parameter}
\usage{
lnorm3_autofit(samples)
}
\value{
fit class from fitdistrplus. Use return$estimate to obtain fitted parameters. See ?fitdistrplus for more info
}
\description{
fit lognormal distribution with a shift parameter
}
