% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runOnSizes.R
\name{runOnSizes}
\alias{runOnSizes}
\title{Measure runtime of an algorithm over a series of generated inputs}
\usage{
runOnSizes(algo, inputGenerator, sizes = c(10, 100), timeLimit = 60,
  algArgs = c(), inputArgs = NULL)
}
\arguments{
\item{algo}{a function to be benchmarked. Expected signature:
`algo(input, algArgs)` where input is returned from the inputGenerator.
Any output will be ignored (to be fixed in the future)}

\item{inputGenerator}{a function with signature: f(size, inputArgs) returning a single input for algo.}

\item{sizes}{vector of input sizes. Passed one by one to the inputGenerator.}

\item{timeLimit}{If that many seconds have passed, next size will not be evaluated. Note this doesn't interrupt
the algorithm}

\item{algArgs}{optional arguments to be passed to algo calls}

\item{inputArgs}{optional arguments to be passed to inputGenerator}
}
\value{
A data frame containing size of the input and measured runtime. May store additional specification, too.
}
\description{
Displays a progress bar of timeLimit consumption, updated every time algorithm returns.
Allows breaking the execution after predefined time.
}
